White Paper and Progress Report
Issuance and settlement methods for yen-denominated private-sector digital currencies, as well as the structure of the forum (74 companies, banks, local governments, and organizations) and use cases by 10 subcommittees. 

DeCurret Inc. (HQ: Chiyoda-ward, Tokyo Japan, President & CEO: Kazuhiro Tokita) is serving as the secretariat of the Digital Currency Forum, a forum to study the practicality of digital currencies in Japan, with the participation of more than 70 companies, banks, local governments, and organizations from a wide range of fields, including Japan's leading banks, retail, transportation, and information and communications, as well as experts and observers from relevant ministries, agencies, and central banks.

The Digital Currency Forum has recently compiled a white paper on the issuance and settlement methods of digital currency on a two-tiered digital currency platform, and a progress report on the activities being conducted by 10 subcommittees to study proof of concept (PoC) and use cases.　

The Digital Currency Forum (hereafter referred to as "the Forum") will release these two reports.